Crate gui

Modules

db 🔒

log 🔒

settings 🔒

Structs

FooterProps 🔒

Properties for the Footer component.

HeaderProps 🔒

Properties for the Header component.

Enums

Tab 🔒

Represents the different tabs in the application.

Constants

CONFIG_FILE 🔒

The path to the configuration file.

FAVICON 🔒

The path to the favicon asset.

TAILWIND_CSS 🔒

The path to the Tailwind CSS asset.

Functions

App 🔒

The main application component. It initializes the application state, including the selected tab, the settings, and the proxy process.

Footer 🔒

Represents the footer of the application. It contains a play button to start the proxy process.

Header 🔒

Represents the header of the application. It contains buttons to switch between different tabs.

main 🔒