Trait gui::Widget

pub trait Widget<E, M>: Handleable<E, M> + Renderable + Object + Debug where
    E: 'static,
    M: 'static, {
    fn type_id(&self) -> TypeId;

    fn data<'c, D>(&self, cap: &'c dyn Cap) -> &'c D
    where
        Self: Sized,
        D: 'static,
    { ... }

    fn data_mut<'c, D>(&self, cap: &'c mut dyn MutCap<E, M>) -> &'c mut D
    where
        Self: Sized,
        D: 'static,
        E: Debug,
        M: Debug,
    { ... }
}

A widget as used by a Ui.

In addition to taking care of Id management and parent-child relationships, the Ui is responsible for dispatching events to widgets and rendering them. Hence, a widget usable for the Ui needs to implement Handleable, Renderable, and Object.

Required methods

fn type_id(&self) -> TypeId

Get the TypeId of self.

Provided methods

fn data<'c, D>(&self, cap: &'c dyn Cap) -> &'c D where
Self: Sized,
D: 'static,

Retrieve a reference to a widget’s data.

Note: This function will panic if the data associated with the object is not of type D.

fn data_mut<'c, D>(&self, cap: &'c mut dyn MutCap<E, M>) -> &'c mut D where
Self: Sized,
D: 'static,
E: Debug,
M: Debug,

Retrieve a mutable reference to a widget’s data.

Note: This function will panic if the data associated with the object is not of type D.

Implementations

impl<E, M> dyn Widget<E, M> where
E: 'static,
M: 'static,

pub fn is<T: Widget<E, M>>(&self) -> bool

Check if the widget is of type T.

pub fn downcast_ref<T: Widget<E, M>>(&self) -> Option<&T>

Downcast the widget reference to type T.

pub fn downcast_mut<T: Widget<E, M>>(&mut self) -> Option<&mut T>

Downcast the widget reference to type T.