Crate gui

A crate containing the basic infrastructure for user interfaces. It strives for being completely agnostic of the underlying system and its rendering machinery as well as event dispatching.

Modules

derive

A module providing custom derive functionality for gui related traits.

Structs

BBox

A bounding box representing the area that a widget may occupy. A bounding box always describes a rectangular area. The origin [x=0, y=0] is typically assumed to reside in the upper left corner of the screen, but it is really up to the individual Renderer to make do with whatever is provided.

Id

An Id uniquely representing a widget.

Ui

A Ui is a container for related widgets.

Traits

Cap

A capability allowing for various widget related operations.

Handleable

A trait representing an object capable of handling events.

Mergeable

A trait for merging two values.

MutCap

A mutable capability allowing for various widget related operations.

Object

An Object represents a first-class entity in a UI.

Renderable

A trait representing a renderable object.

Renderer

An abstraction for objects used for rendering widgets.

Widget

A widget as used by a Ui.